Remove uint32_t-based file size limits
- Change loader abstraction to use 64-bit ints for indices
- Make BP128 loaders have a secondary idx_offsets field to handle when
  the main idx overflows
- Make StoredMatrix use a uint64_t for col_ptr
- Make StoredFragments use a uint64_t for chr_ptr

+++ src/arrayIO/array_interfaces.cpp
@@ -3,10 +3,10 @@
 namespace BPCells {
 
 VecStringReader::VecStringReader(std::vector<std::string> data) : data(data) {}
-const char *VecStringReader::get(uint32_t idx) const {
+const char *VecStringReader::get(uint64_t idx) const {
     if (idx < data.size()) return data[idx].c_str();
     return NULL;
 }
-uint32_t VecStringReader::size() const { return data.size(); }
+uint64_t VecStringReader::size() const { return data.size(); }
 
 } // end namespace BPCells
